fix: resolve CodeQL warnings in signal_detector and nav2z planners
- signal_detector.cpp: store demangled type string before calling c_str()
  to prevent dangling pointer (cpp/use-of-string-after-lifetime-ends)
- forward_local_planner, backward_local_planner: cast float velocity
  components to double before multiplication to avoid float precision
  loss before widening (cpp/integer-multiplication-cast-to-long)
- cb_spiral_motion: same float-to-double cast for linearAceleration * dt

+++ smacc2_client_library/cl_nav2z/cl_nav2z/src/cl_nav2z/client_behaviors/cb_spiral_motion.cpp
@@ -93,7 +93,7 @@ void CbSpiralMotion::onEntry()
   {
     auto current_time = getNode()->now() - start_time;
 
-    cmd_vel.linear.x += linearAceleration * dt;
+    cmd_vel.linear.x += static_cast<double>(linearAceleration) * dt;
     if (cmd_vel.linear.x > maxLinearVelocity)
     {
       cmd_vel.linear.x = maxLinearVelocity;
